+++ cmds/command_search.go
@@ -49,13 +49,15 @@ func searchFunc(cmd *Command, _ ...string) error {
 
 	var handler handlers.SearchHandler
 	allHandlers := map[string]handlers.Handler{
-		consts.TextQuery: handlers.SearchTextQueryHandler{},
-		consts.IDLookup:  handlers.SearchIDLookupHandler{},
+		consts.TextQuery:  handlers.SearchTextQueryHandler{},
+		consts.IDLookup:   handlers.SearchIDLookupHandler{},
+		consts.ExactQuery: handlers.SearchExactQueryHandler{},
+		consts.Trending:   handlers.SearchTrendingHandler{},
 	}
 	handler, err := cmd.common.GetHandlerForMap(options.Action, allHandlers)
 
 	if err != nil {
-		cmd.common.GenActionsUsage(cmd.Name, []string{consts.TextQuery, consts.IDLookup})
+		cmd.common.GenActionsUsage(cmd.Name, []string{consts.TextQuery, consts.IDLookup, consts.ExactQuery, consts.Trending})
 		return nil
 	}